Revert "Use fread instead of loop of fgetc"
I was saving about 0.01s at runtime, but it made valgrind take ages, like
several minutes.

The main reason for using C at all is that I want a fast feedback loop
during development, so it's better to be 0.01s slower at runtime and
be able to run Valgrind quickly.

This reverts commit 46351aed3738478246f5ba741448c079886a8a23.