Reverted back to Haskell implementation
I was trying doing it again from scratch. The reason was that I was tired
of the previous approach of just dealing with bug after bug.

The idea was that by starting over I could reduce the number of bugs by
adding a lot of formatted tests at the start.

However I'm not sure this actually helps any more. Because having a lot of formatted tests is going to result in a huge batch of failures and bugs as
soon as the formatter doesn't fail totally on those files.

So I'm coming to the conclusion that the best approach is the one I used
before, which was basic test-driven development: write a minimal test,
then a minimal implementation, and then repeat.