Fixed the float parser
I had made a parser for floats with both a dot and an exponent, such as
1.0e0.

The formatter was failing when it was trying to parse the final `\n\0'.

The issue was that the new float parser was run after attempting the simple
float parser, for things like 1.0. So the simple parser was consuming the
1.0 and stopping at the e. And my new parser for 1.0e0 was never being
run.

And then when the formatter tried to parse the trailing '\n\0' it was
hitting the 'e' and failing.

The solution was to move the new parser for 1.0e0 type floats above the
one for simple floats.