Trivial implementation of format library in Go
The previous implementation was in C and the one before that was in Haskell.

The reason for using Haskell was that I was used to it, especially the
parser combinator libraries. However I realised after a while that it is
easy to do parser combinators using ordinary functions in an imperative
language, so that wasn't really a good reason for using Haskell any more.

One pain point with Haskell was the slow compile times. I found that the
fastest I could do an incremental compilation was about 7 seconds. This was
with the compiler optimisations turned off. It doesn't sound much but I was
doing test-driven development and was spending a lot of the time waiting
for those 7 seconds.

Another problem with Haskell is that it's hard to cross compile. This is
a nuisance because I want to be able to support at least Mac, Linux and
Windows.

It's also time-consuming to build from source if you haven't got Haskell
already. So it's asking a lot to expect users to build Haskell code from
source.

Another problem with Haskell is that there are a lot of major versions of
GHC. This means that you have to have a lot of compilers around in Haskell
development. Or you can keep upgrading.

Then I was trying C. The main reasons were very fast build times and very
fast run times. I also like the language. It's also much easier to build it
from source and to cross compile.

I was really enjoying working in C. I could compile and run about 3000
tests in about 0.2s. This makes a huge difference when you are working with
test-driven development.

The big problem was that I was often running into memory issues. One way I
would notice these was by the tests breaking on some trivial change. For
example reordering the arguments in a function or marking a function as
static. I would also get segmentation faults.

The only way I could find to debug these issues was to run the tests in
Valgrind. Valgrind is a great tool and this method worked really well.
However I was having to run Valgrind really often to check if I had
introduced a memory issue. And my test suite was taking around 6 minutes
to run in Valgrind.

So effectively when developing in C I had a very slow feedback loop, much
too slow to do test-driven development.

I started shopping around for a different language. The requirements were:

- fast at runtime
- fast to do compilation + tests + memory safety checks
- easy to cross compile or easy for users to build from source
- stable

I looked at a lot of languages.

I thought scripting languages like Python, Ruby or Javascript would be
too slow, and also require that users have the right version of the
runtime.

Rust has a really slow compiler.

I'm not very familiar with OCaml but it seems it is difficult to
cross-compile and I don't think I can expect my users to install an OCaml
toolchain.

Zig looks great but it isn't stable yet. I might also find I need to run
Valgrind as its memory safety isn't very strong.

I had a look at C#. It has a lot of nice properties. It is one of the
fastest languages with automatic memory management. The benchmarks in the
benchmarks game website have it consistently faster than Go. However when
I looked at the implementations the C# ones were dropping down into unsafe
code, so I might as well carry on using C.

It seems a bit much to expect users to install dot net, but I found that
you can bundle C# apps into a self-contained binary, although it is rather
large.

I also had a look at Nim. It looks like a great language and would allow
me to easily cross compile. It seems very performant. The main thing that
put me off was the lack of stability. There was a version 2.0, which must
have broken a lot of 1.0 code. I want to avoid the Haskell situation of
having dependencies on a particular compiler version.

I considered Roc. I think it's going to be a brilliant language one day,
and is probably performant enough. However it is not at all stable yet and
I don't want to have to do large upgrades.

The main issue with Go is that it is significantly slower than C. It's hard
to be precise without implementing the whole thing in both C and Go and
comparing, but a ballpark figure from browsing toy benchmarks online would be
that Go programs are several times slower than C ones, maybe 2 to 5 times
slower.

However Go performs very well compared to other languages with automatic
memory management.

Go ticks off the other requirements. It is good at cross compiling, it
compiles really fast, there are no memory leaks, it's easy to deploy, it's stable.

By switching from C to Go I am getting a much shorter development feedback loop
in exchange for less performance at run time.