Count the number of failed tests
This allows the test runner to return a success or failure status code.

This means that the test script can quit early if a test fails.

This means that I don't have to have repeated failure messages, which is
quite annoying.

The reason for running the test suite multiple times is so that different
sanitizers can be run.