Redirect stdout from the second test run to /dev/null
It prevents having two lots of test results.

Previously I was hiding the results of the first test run, but the problem
was that when the first test run is unsuccessful this means that debug
print statements are hidden.

When the first teset run fails the second run doesn't happen.