Got rid of function for calculating the total comment size
This function was used by the writer to see if it needed to add a
newline and indent after a comment. It was necessary because if the
line comment parser failed it would insert an empty literal node into
the tree.

I simplified the comment parser so that empty comments are not inserted
into the tree at all. This means that the writer can just check if the
whitespace parent has any children instead of the size calculation.