It no longer creates a temporary file
Previously the formatted code was written to a file after each top level
definition. It would write the formatted code to a temporary file and then
rename the temporary file to the main file when finished.

But now the formatter reads all the code for a file into memory up front so
there is no need for the temporary file.