Formatter for empty multiline block comments
This one is a bit tricky because elm-format behaves strangely with it.

1. Elm-format leaves {--} as it is
2. Elm-format changes {-  -} to {- -}
3. Elm-format changes {-\n-} to {-  -}

So I just change any empty block comment to {--} :shrug: At least it's
consistent.

There isn't a test for this because it's confusing to have tests that are
different to elm-format.

But I couldn't leave it as it was because it was failing on empty block
comments.

Hopefully this corner case won't come up much in real code.