Preserve blank line at start of hanging block comment
This is for block comments like this:

{-

   a

-}

These are allowed by elm-format. But trumat was removing the initial
blank line. The reason was that there was a whitespace consumer at
the first step of the parser after the opening.

I also had to fix the writer function because it wasn't checking if the
first line was empty, so was inserting some spaces.