Got rid of error enum
I was getting in a mess with handling all the errors. I had a huge enum
for all the places where errors can happen. This is a lot in a complicated
parser like this one.

The only benefit of these errors in for development. But it's difficult
to decide when to bubble the errors up or when to replace them with new
ones.

In the end I decided to just return -1 for all the errors and use the
debugger and printf to fix issues.