Format triple quoted strings without modifying the source buffer
I'm gradually replacing the parsers that modify the source buffer because
I want to use it to compare to the formatted output.

Currently I have a separate buffer containing the original source before
it is modified.

The only places where the source is modified is for hex literals which
need to be uppercased. These are not that common and it seems wasteful
to duplicate the source buffer for this.

The places where hex literal are used are:
- integer literals
- unicode literals in triple quoted strings
- unicode literals in normal strings