Don't consume whitespace in function argument parser if it fails
Previously the argument parser was consuming the initial whitespace
even if it failed. This meant that it would think that some function calls
were multiline when they weren't.

The bug was masked by using ROW[I-1] instead of ROW[I]. I had put this in
because I mistakenly thought that changing the parser index to start at 0
instead of -1 meant I needed to reduce the ROW index.